my commit for 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,24 +7,24 @@
 int main(void)
 {
 	int a = 0;
-	int b = 1;
-	int c = 2;
+	long b = 1;
+	long c = 2;
 
 	while (a < 50)
 	{
 		if (a == 0)
 		{
-			printf("%d", b);
+			printf("%ld", b);
 		}
 		else if (a == 1)
 		{
-			printf(", %d", c);
+			printf(", %ld", c);
 		}
 		else
 		{
 			c += b;
 			b = c - b;
-			printf(", %d", c);
+			printf(", %ld", c);
 		}
 		++a;
 	}
